SilverTier/requirements.txt
groq>=0.9.0
python-dotenv>=1.0.0
imapclient>=2.3.1
watchdog>=3.0.0
requests>=2.31.0
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.27.0
aiosmtplib>=3.0.0
pydantic>=2.7.0
